Move from cleanup to storage
